day 2 complete
Adds the solution to day 2. I liked my solution to the first part since
the cross product neatly solves the rock paper scissors game, however my
brute force approach to the second part was maybe a bit unnecessary. I
wonder if there is any nice mathematical way to come up with a better
solution without using if/else statements. Also, the brute force method
requires argument checking of tuples, which for this small example is
fine but may not be computationally optimal.